fix: goal suggestion funtionality not working
