+++ src/components/ConfigGoal/ConfigGoal.scss
@@ -32,6 +32,7 @@
     align-items: center;
     justify-content: center;
     padding: 24px 12px;
+    border-radius: 28px;
   }
   .header-title > input {
     font-size: 1.5rem;
