+++ src/components/ConfigGoal/components/AutoComplete.scss
@@ -2,14 +2,19 @@
   position: relative;
   display: flex;
   align-items: center;
+  padding: 0;
+  font-size: inherit;
+  font-family: inherit;
 
   input {
     background: transparent;
+    border: none;
     font-size: 1.3rem;
     outline: none;
     padding-right: 0;
     margin: 0;
     width: auto;
+    position: relative;
     z-index: 2;
   }
 
@@ -33,7 +38,7 @@
     pointer-events: auto;
     white-space: nowrap;
     user-select: none;
-    z-index: 1;
+    z-index: 100;
     padding: 5px 0 0 12px;
     margin: 0;
     font-family: inherit;
